merging Fed hicop that has all the features raised and implemented during the Hackathon into dev.

+++ src/main/java/org/lamisplus/modules/pmtct/domain/entity/Infant.java
@@ -25,6 +25,6 @@ public class Infant extends PMTCTTransactionalEntity implements Serializable, Pe
     private LocalDate nextAppointmentDate;
     private Integer defaultDays;
     private String motherPersonUuid;
-    private Long bodyWeight;
+    private Double bodyWeight;
     private String ctxStatus;
    }
